+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ### Changed
 
-## Unreleased
+**Release 0.1.5** - 2020-09-25
+
+-   Fixed bug for init_app method from regression bug  [Issue #99](https://github.com/joegasewicz/flask-file-upload/issues/99)
+-   Fixed bug for init_app method from regression bug  [Issue #97](https://github.com/joegasewicz/flask-file-upload/issues/97)